mac-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.68.2
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.0.1
pymongo==4.13.2
